+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.2.0
 
 * Added top level crate
+* Reorganize nsm-driver and nsm-io into top level crate
 * Changed authors in Cargo.toml